Move counters to using double instead of float.

+++ checker/tracking/include/CheckerTypes.h
@@ -99,8 +99,8 @@ namespace Checker {
     std::size_t m_nclones = 0;
     std::size_t m_nevents = 0;
     float m_effperevt = 0.f;
-    float m_hitpur = 0.f;
-    float m_hiteff = 0.f;
+    double m_hitpur = 0.0;
+    double m_hiteff = 0.0;
     std::size_t m_naccept_per_event = 0;
     std::size_t m_nfound_per_event = 0;
     std::size_t m_nclones_per_event = 0;
